+++ test/opf-var.jl
@@ -343,6 +343,16 @@ end
         end
     end
 
+    @testset "test soc opf" begin
+        @testset "5-bus uc case" begin
+            result = PowerModels._run_ucopf("../test/data/matpower/case5_uc.m", SOCWRPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 15057.09; atol = 1e0)
+            @test isapprox(result["solution"]["gen"]["4"]["gen_status"], 0.0, atol=1e-6)
+        end
+    end
+
     @testset "test dc opf" begin
         @testset "5-bus uc case" begin
             result = PowerModels._run_ucopf("../test/data/matpower/case5_uc.m", DCPPowerModel, cbc_solver)
@@ -353,14 +363,43 @@ end
         end
     end
 
+
     @testset "test ac opf" begin
         @testset "5-bus uc storage case" begin
             result = PowerModels._run_ucopf("../test/data/matpower/case5_uc_strg.m", ACPPowerModel, juniper_solver)
 
             @test result["termination_status"] == LOCALLY_SOLVED
             @test isapprox(result["objective"], 17740.9; atol = 1e0)
+
+            @test isapprox(result["solution"]["gen"]["4"]["gen_status"], 0.0, atol=1e-6)
             @test isapprox(result["solution"]["storage"]["1"]["status"], 1.0, atol=1e-6)
             @test isapprox(result["solution"]["storage"]["2"]["status"], 0.0, atol=1e-6)
+        end
+    end
+
+    @testset "test soc opf" begin
+        @testset "5-bus uc storage case" begin
+            result = PowerModels._run_ucopf("../test/data/matpower/case5_uc_strg.m", SOCWRPowerModel, juniper_solver)
+
+            @test result["termination_status"] == LOCALLY_SOLVED
+            @test isapprox(result["objective"], 14525.0; atol = 1e0)
+
+            @test isapprox(result["solution"]["gen"]["4"]["gen_status"], 0.0, atol=1e-6)
+            @test isapprox(result["solution"]["storage"]["1"]["status"], 1.0, atol=1e-6)
+            @test isapprox(result["solution"]["storage"]["2"]["status"], 0.0, atol=1e-6)
+        end
+    end
+
+    @testset "test dc opf" begin
+        @testset "5-bus uc storage case" begin
+            result = PowerModels._run_ucopf("../test/data/matpower/case5_uc_strg.m", DCPPowerModel, cbc_solver)
+
+            @test result["termination_status"] == OPTIMAL
+            @test isapprox(result["objective"], 16833.2; atol = 1e0)
+
+            @test isapprox(result["solution"]["gen"]["4"]["gen_status"], 0.0, atol=1e-6)
+            @test isapprox(result["solution"]["storage"]["1"]["status"], 1.0, atol=1e-6)
+            @test isapprox(result["solution"]["storage"]["2"]["status"], 1.0, atol=1e-6)
         end
     end